-added acr dialogs -show current weather on nearby events page

+++ Nearby/Nearby/Models/WatherResult.cs
@@ -40,16 +40,16 @@ namespace Nearby.Models
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
-        public int temp_min { get; set; }
-        public int temp_max { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
+        public double temp_min { get; set; }
+        public double temp_max { get; set; }
     }
 
     public class Wind
     {
-        public int speed { get; set; }
-        public int deg { get; set; }
+        public double speed { get; set; }
+        public double deg { get; set; }
     }
 
     public class Clouds
